+++ Baboon.Angular.Common/common/authentication/authentication.module.js
@@ -1,8 +1,16 @@
-(function(){
+(function () {
 	'use strict';
-
+			/*
+			
+			Choose either basic or JWT authentication.
+			
+			Note that the 'authentication service' will have different features exposed depending on which authentication you choose.
+			
+			*/
 	angular.module('angular-baboon.common.authentication', [
-            'angular-baboon.common.heartbeat'
-		]);
+
+		'angular-baboon.common.authentication.jwt',
+		//'angular-baboon.common.authentication.basic'
+	]);
 
 })();