+++ Baboon.Angular.Common/common/authentication/basic/basic.config.js
@@ -1,0 +1,9 @@
+(function () {
+    'use strict';
+
+    angular.module('angular-baboon.common.authentication.basic')
+        .config(function ($httpProvider) {
+            $httpProvider.interceptors.push('basicAuthHttpInterceptor');
+        });
+
+})();