+++ Baboon.Angular.Common/common/authentication/basic/basic.module.js
@@ -1,0 +1,10 @@
+(function () {
+	'use strict';
+
+	angular.module('angular-baboon.common.authentication.basic', [
+		'angular-baboon.common.authentication.basic.factory',
+		'angular-baboon.common.authentication.basic.interceptor',
+		'angular-baboon.common.authentication.basic.heartbeat'
+	]);
+
+})();