+++ Baboon.Angular.Common/common/authentication/basic/services/authHttpInterceptorFactory.js
@@ -2,8 +2,8 @@
     'use strict';
     // factory
     angular
-        .module('angular-baboon.common.authentication')
-        .factory('authHttpInterceptor', ['$log', '$q', '$injector', '$rootScope', function ($log, $q, $injector, $rootScope) {
+        .module('angular-baboon.common.authentication.basic.interceptor', [])
+        .factory('basicAuthHttpInterceptor', ['$log', '$q', '$injector', '$rootScope', function ($log, $q, $injector, $rootScope) {
             var authentication;
             return {
                 'requestError': function (rejection) {
@@ -15,7 +15,7 @@
                     return $q.reject(rejection);
                 },
                 'responseError': function (rejection) {
-					if (rejection.status === 401) {
+                    if (rejection.status === 401) {
                         authentication = $injector.get('authentication');
                         authentication.unauthorizedRequest();
                         $rootScope.$broadcast('unauthorized-request');