+++ Baboon.Angular.Common/common/authentication/heartbeat/heartbeat.config.js
@@ -1,9 +1,0 @@
-(function () {
-    'use strict';
-
-    angular.module('angular-baboon.common.heartbeat')
-        .config(function () {
-            
-        });
-
-})();