+++ Baboon.Angular.Common/common/authentication/heartbeat/heartbeat.module.js
@@ -1,7 +1,0 @@
-(function(){
-	'use strict';
-
-	angular.module('angular-baboon.common.heartbeat', [
-		]);
-
-})();