+++ Baboon.Angular.Common/common/authentication/jwt/jwt.config.js
@@ -1,0 +1,9 @@
+(function () {
+    'use strict';
+
+    angular.module('angular-baboon.common.authentication.jwt')
+        .config(function ($httpProvider) {
+            $httpProvider.interceptors.push('jwtAuthInterceptor');
+        });
+
+})();