+++ Baboon.Angular.Common/common/authentication/jwt/jwt.module.js
@@ -1,0 +1,9 @@
+(function () {
+	'use strict';
+
+	angular.module('angular-baboon.common.authentication.jwt', [
+		'angular-baboon.common.authentication.jwt.factory',
+		'angular-baboon.common.jwt-authentication.interceptor'
+	]);
+
+})();